requirements.txt
pandas
matplotlib
edgartools # MIT License
numpy
